Implement instructions with variable register size
Loading operands is a lot more complex since they vary depending
on the CPU state and which register is accessed.
Putting the information into an Operand struct hides some of that
complexity, but should still be easy to optimize by the compiler,
since most instructions will just pass in constants.