Generic variable register size
The A, X and Y registers can be u8 or u16. I want to implement the
instruction logic in a generic way for both sizes. To do so, I am
adding a RegisterValue trait that has all the operations
needed to implement them, and is implemented for both u8 and u16.
This should make the instruction implementation repetition free.