Implement a couple more address modes
Now runs until instruction 229526 of the test. It looks like we
got all the instructions for the general framework of the tests,
so now we can focus on the actual instruction being tested (adc).
Address mode code is getting annoyingly repetitive. There is
a lot of room for improvement there.