Implement CPUORA and CPUPHL tests
The stack tests were a bit tricky. Not to implement but to re-create
the BSNES trace disassembly, it's a bit inconsistent, at least
from an outside perspective.
I updated the cpu_tests.rs code to skip some comparisons in the traces
so that I do not have to spend time on re-creating the exact format that
BSNES generates.