Add PpuTimer to calculate h, v location of PPU
Since BSNES will output timing information only in the form of PPU
h, v location, we need to calculate those in order to validate cpu
instruction timing in cpu_tests.rs.

Unfortunately the timing is not straight-forward as dots have variable
durations, and dram refresh will pause the CPU for a few cycles.

The timing is validated with a log from bsnes executing no-ops.