Add ReadOrPeekWrapper to unify operand loading
Loading operands is complex and full of little nits that need careful
testing. But there need to be two implementations of the same logic,
one that loads operands for disassembly, using only peeks and immutable
access to the cpu.
And another one used during execution that will be able to mutate
the cpu.
ReadOrPeekWrapper is a bit of ugly boilerplate to make sure we
only need one implementation of the tricky logic.