spc700: Let instructions decode operands
For instructions with multiple operands, loading both operands
upfront will result in the wrong read order on the bus.
To fix this, just pass OperandDef to the instructions and decode
operands there.
This is probably overkill and can be simplified, but let's keep
this flexible until all opcodes are implemented.